% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionSetDF.R
\docType{class}
\name{expressionSetDF}
\alias{expressionSetDF}
\alias{expressionSetDF-class}
\title{Clase expressionSetDF}
\description{
Esta clase representa un conjunto de datos de expresión genética con campos exprs,
phenoData y featureData, donde exprs es un data.frame.
}
\section{Slots}{

\describe{
\item{\code{exprs}}{Data frame con datos de expresión genética.}

\item{\code{phenoData}}{Data frame con información fenotípica.}

\item{\code{featureData}}{Data frame con información de características.}
}}

