% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionSetDF.R
\name{MultMerge2}
\alias{MultMerge2}
\title{MultMerge2}
\usage{
MultMerge2(lst, all.x = TRUE, all.y = TRUE, by = NULL)
}
\arguments{
\item{lst}{Una lista de data frames que se fusionarán.}

\item{all.x}{Si TRUE, incluir todas las filas de los data frames en la lista original.}

\item{all.y}{Si TRUE, incluir todas las filas de los data frames en la lista nueva.}

\item{by}{Nombre de la columna por la que se unirán los data frames (opcional).}
}
\value{
Un nuevo data frame fusionado.
}
\description{
Combina varios data frames en un único data frame, 
manteniendo todas las filas y columnas únicas de cada uno.
}
\examples{
df1 <- data.frame(matrix(rnorm(20), nrow=10))
df2 <- data.frame(min=letters[1:8], may=LETTERS[1:8])
dfList <- list(df1, df2)
MultMerge2(dfList)

}
